% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grafico_temperatura_mensual.R
\name{grafico_temperatura_mensual}
\alias{grafico_temperatura_mensual}
\title{Gráfico de temperatura promedio mensual por estación}
\usage{
grafico_temperatura_mensual(
  datos,
  colores = "darkgreen",
  titulo = "Temperatura"
)
}
\arguments{
\item{datos}{Un data frame que contenga al menos las columnas \code{fecha}, \code{id}
y \code{temperatura_abrigo_150cm}.}

\item{colores}{Un vector con los colores a utilizar para las estaciones.
Por defecto \code{"darkgreen"}.}

\item{titulo}{Título del gráfico. Por defecto \code{"Temperatura"}.}
}
\value{
Un objeto \code{ggplot} con el gráfico generado.
}
\description{
Gráfico de temperatura promedio mensual por estación
}
\examples{
grafico_temperatura_mensual(datos)
grafico_temperatura_mensual(datos, colores = c("red", "blue"), titulo = "Temperatura mensual promedio")

}
