% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leer_estacion.R
\name{leer_estacion}
\alias{leer_estacion}
\title{Leer y descargar datos de una estación meteorológica}
\usage{
leer_estacion(id_estacion, ruta_archivo)
}
\arguments{
\item{id_estacion}{carácter, identificador de la estación.}

\item{ruta_archivo}{carácter, ruta donde se guardará o leerá el archivo CSV.}
}
\value{
Un data frame con los datos de la estación meteorológica.
}
\description{
Esta función descarga (si es necesario) y lee los datos de una estación meteorológica
a partir de su identificador. Si el archivo ya existe en la ruta indicada, se omite
la descarga y se leen los datos directamente desde el disco.
}
\examples{
datos <- leer_estacion("NH0437", "datos/NH0437.csv")
head(datos)

}
