% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_resumen_temperatura.R
\name{tabla_resumen_temperatura}
\alias{tabla_resumen_temperatura}
\title{Tabla resumen de temperatura por estación}
\usage{
tabla_resumen_temperatura(...)
}
\arguments{
\item{...}{Uno o más data frames que contengan las columnas \code{id} y
\code{temperatura_abrigo_150cm}.}
}
\value{
Un data frame con las columnas:
\itemize{
\item \code{id}: identificador de la estación.
\item \code{variable}: nombre de la variable analizada.
\item \code{media}: temperatura promedio.
\item \code{mediana}: temperatura mediana.
\item \code{minimo}: temperatura mínima.
\item \code{maximo}: temperatura máxima.
}
}
\description{
Esta función combina los datos de una o más estaciones meteorológicas y genera
una tabla resumen con estadísticas básicas de la variable
\code{temperatura_abrigo_150cm} para cada estación.
}
\examples{
tabla_resumen_temperatura(estacion1, estacion2)

}
